Finish database features, user login/registration
This commit implements database features to the project. This allows
data to be stored in a database for easier access. This commit also
implements user login and registration. This allows a user to log in
with a username and password. The password is stored in the database
hashed to allow for enhanced security. This is done using Spring
Security Crypto.